+++ packages/studio-base/src/panels/Image/components/ImageCanvas.tsx
@@ -24,7 +24,7 @@ import Rpc from "@foxglove/studio-base/util/Rpc";
 import WebWorkerManager from "@foxglove/studio-base/util/WebWorkerManager";
 
 import { renderImage } from "../lib/renderImage";
-import { Config, SaveImagePanelConfig } from "../types";
+import { Config } from "../types";
 import type {
   Dimensions,
   PixelData,
@@ -42,7 +42,7 @@ type Props = {
   image?: NormalizedImageMessage;
   rawMarkerData: RawMarkerData;
   config: Config;
-  saveConfig: SaveImagePanelConfig;
+  saveConfig: (config: Partial<Config>) => void;
   onStartRenderImage: () => OnFinishRenderImage;
   renderInMainThread?: boolean;
   setActivePixelData: (data: PixelData | undefined) => void;
